# docs/openapi/orbit.yaml
openapi: 3.1.0
info:
  title: Orbit Placement CRM API
  version: 0.1.0
  description: Authenticated API for Orbit recruitment and placement operations.
servers:
  - url: http://localhost:3101/api/v1
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: orbit_session
  schemas:
...
    BulkImportRequest:
      type: object
      required: [csv]
      properties:
        csv: { type: string, minLength: 1, maxLength: 5242880 }
    BulkImportResult:
      type: object
      required: [imported, failed, errors]
      properties:
        imported: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
        errors: { type: array, items: { type: object, required: [row, message], properties: { row: { type: integer, minimum: 1 }, message: { type: string } } } }
...
paths:
...
  /imports/candidates: { post: { summary: Bulk import candidates, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BulkImportRequest' } } } }, responses: { '200': { description: Import result, content: { application/json: { schema: { $ref: '#/components/schemas/BulkImportResult' } } } } } } }
  /imports/leads: { post: { summary: Bulk import leads, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BulkImportRequest' } } } }, responses: { '200': { description: Import result, content: { application/json: { schema: { $ref: '#/components/schemas/BulkImportResult' } } } } } } }
...